Always create 8 'guixbuilder' user accounts
The number of concurrent offload builds is limited by the number of
'guixbuilder' accounts. Previously this was set to be the number of
cores on the machine building the AMI, which could be as low as 2. It
is often desired to have at least 3 concurrent builds, although since
each build can attempt to utilize all cores it is rare to need
additional concurrent builds. A rough, untested rule-of-thumb might be
to use log2(# cores) concurrent builds, and since the maximum number of
cores today is 256 the number of accounts is set to 8.