Run multiple concurrent jobs from guix-daemon
The number of concurrent jobs is set to one plus the binary logarithm of
the number of CPU cores. In addition to concurrent builds, this allows
multiple substitutes to be downloaded concurrently on a multi-core system.

Cores | Jobs
----- | ----
    1 |    1
    2 |    2
    4 |    3
    8 |    4
   16 |    5
   32 |    6
   64 |    7
  128 |    8

The number of build users is also increased from 8 to 16 to future proof
against core count increases since Guix fails when starting a build if
the necessary guixbuilder user has not been created.