Bootstrap Guix from source
Adds an option to the Packer configuration to disable use of Guix
substitute packages, meaning that after downloading and installing the
binary tarball all packages will be compiled from source.

Several bootstrap packages cannot be compiled on an EBS volume either
because of the NVME interface or block number, so instead a tmpfs mount
is used and the /gnu and /var/guix directories copied back-and-forth.

For updating bootstrap builds on a future instance (which would be
expected after a merge of core-updates) it should be possible to
similarly configure a second offload machine which would copy over the
minimum number of packages.